Add trigger detection and fix igniter state setting.
1) Add detection of trigger going HIGH on pin 2 and turn on igniter for 1s. Not
that currently the collision trigger will keep resetting the timer as long as
pin 2 is HIGH, while the controller only resets the timer when the controller
changes igniter state to on. Also, the state/timeout is shared between both
causes, and can be reset by either. We may want to change some of these
behaviors. (I'd vote swapping the two trigger behaviors but leaving the state
shared.) Resolves #15.

2) Fixed a bug in igniter state differencing where it would fail to update the
current state when the igniter setting was changed.