Allow variable-length messages; raised protocol version to 2 due to variable-length messages; added remote configuration of some variables; update protocol documentation to include remote configuartion message.
For remote configuration, all configurable variables are
stacked in a single message made of up 16-bit short integers.
If an integer is 0xFFFF, which they are if the corresponiding
constant in constants.py is set to None, then the value is
set to its default on the rfduino side.  See protocol
documentation for more information on the variable
selection/ordering.